+++ Kernel/System/Console/Command/QueuesGroupsRoles/Download.pm
@@ -42,9 +42,7 @@ use utf8;
 sub Configure {
     my ( $Self, %Param ) = @_;
 
-    $Self->Description(
-        'Download of OTRS queues-groups-roles schema into CSV-file, Copyright (c) 2006-2015 c.a.p.e. IT GmbH, http//www.cape-it.de/'
-    );
+    $Self->Description('Download of OTRS queues-groups-roles schema into CSV-file');
     my @HeadlineKeys = qw{
         SalutationID SignatureID FollowUpID FollowUpLock UnlockTimeout
         FirstResponseTime FirstResponseNotify UpdateTime UpdateNotify